source/kafka_consumer/requirements.txt
confluent-kafka==2.3.0
kafka-python==2.0.2
python-dotenv==1.0.0
